Move HTTPClientSpy to its own file
